rename file to kebab case
