+++ typescript/common-resolvers/src/image/get-workspace-id-of-dataset.ts
